Separated datasource and delegate to their own files.
